// src/features/valuationCompletedLeads/types.ts
export interface LeadListStatuswiseRespDataRecord {
  Id: number;
  LeadUId: string;
  CustomerName: string;
  CustomerMobileNo: string;
  VehicleType: string;
  Vehicle: string;
  LeadType: number;
  LeadTypeName: string;
  StatusId: number;
  Status: string;
  AddedBy: number;
  AddedByDate: string;
  UpdatedBy: number;
  UpdatedByDate: string;
  RegNo: string;
  ChassisNo: string;
  EngineNo: string;
  ProspectNo: string;
  LeadRemark: string;
  ViewUrl: string;
  DownLoadUrl: string;
  QcUpdateDate?: string;
  PriceUpdateDate?: string;
}
export interface LeadListStatusWiseResp {
  Error: string;
  MESSAGE: string;
  DataRecord: LeadListStatuswiseRespDataRecord[];
}

export interface AppLeadDaybookDataRecord {
  lastmonth: number;
  thismonth: number;
  Today: number;
}

export interface AppLeadDaybookResp {
  Error: string;
  MESSAGE: string;
  DataRecord: AppLeadDaybookDataRecord[];
}